+++ src/dipdup/config/__init__.py
@@ -536,7 +536,6 @@ class AdvancedConfig:
     :param scheduler: `apscheduler` scheduler config.
     :param postpone_jobs: Do not start job scheduler until all indexes reach the realtime state.
     :param early_realtime: Establish realtime connection and start collecting messages while sync is in progress (faster, but consumes more RAM).
-    :param skip_version_check: Disable warning about running unstable or out-of-date DipDup version.
     :param rollback_depth: A number of levels to keep for rollback.
     :param decimal_precision: Overwrite precision if it's not guessed correctly based on project models.
     :param unsafe_sqlite: Disable journaling and data integrity checks. Use only for testing.
@@ -547,7 +546,6 @@ class AdvancedConfig:
     scheduler: dict[str, Any] | None = None
     postpone_jobs: bool = False
     early_realtime: bool = False
-    skip_version_check: bool = False
     rollback_depth: int | None = None
     decimal_precision: int | None = None
     unsafe_sqlite: bool = False